feat(oidc): jwt response for introspection
This implements the standard for JWT encoded and signed responses from the introspection endpoint. This has been implemented as per the IETF draft https://www.ietf.org/archive/id/draft-ietf-oauth-jwt-introspection-response-12.html and as it is a draft and that draft is currently expired it should be noted that this implementation may be removed or changed without any regard for breaking changes. While these factors all point in the direction of this may never receive ratification the IANA has accepted registration of the metadata parameters for this specification.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>